requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
pandas
gspread
google-auth
